fix: 第三轮 review - 修复 5 个 High bug
1. ProfileUpdateTool: 使用原子 updatePreferences 防止并发丢失更新
2. ScenarioRouter: 使用浮点数除法避免整数截断
3. WikiSearchTool: 添加 null userId 检查
4. WikiWriteTool: 添加 null userId/sessionId 检查
5. NvcPracticeController: sendMessage/sendMessageStream 添加 @Valid

测试: 全部通过
编译: 通过

+++ app/src/main/java/nvc/guide/modules/nvcpractice/router/ScenarioRouter.java
@@ -78,6 +78,6 @@ public class ScenarioRouter implements ModeRouter {
         if (eval.getFeelingScore() != null) { sum += eval.getFeelingScore(); count++; }
         if (eval.getNeedScore() != null) { sum += eval.getNeedScore(); count++; }
         if (eval.getRequestScore() != null) { sum += eval.getRequestScore(); count++; }
-        return count > 0 && (sum / count) < LOW_SCORE_THRESHOLD;
+        return count > 0 && ((double) sum / count) < LOW_SCORE_THRESHOLD;
     }
 }
